requirements.txt
python-telegram-bot==21.7
supabase==2.4.0
python-dotenv==1.0.0
schedule==1.2.0